Add fire_replaced_class_double.
fire_replaced_class_double is like fire_class_double, plus it temporarily replaces the named constant with itself.  This is handy when the code-under-test references the named constant.  The double also registers itself with rspec-mocks space to ensure it gets reset after each example.  When it is reset it ensures all constants are as they originally were.